\name{ddpca-package}
\alias{ddpca-package}
\alias{ddpca}
\docType{package}
\title{
\packageTitle{ddpca}
}
\description{
\packageDescription{ddpca}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ddpca}
\packageIndices{ddpca}
This package contains \code{\link{DDPCA_nonconvex}} and \code{\link{DDPCA_convex}} function, which decomposes a positive semidefinite matrix into a low rank component, and a diagonally dominant component using either nonconvex approach or convex approach. 
}
\author{
\packageAuthor{ddpca}

Maintainer: \packageMaintainer{ddpca}
}

\references{
Ke, Z., Xue, L. and Yang, F., 2019. Diagonally Dominant Principal Component Analysis. Journal of Computational and Graphic Statistics, under review.
}

\note{
Please cite the reference paper to cite this \R package.
}

\keyword{ package }
\keyword{ ddpca }


